+++ src/ui/tokens/space.ts
@@ -6,7 +6,7 @@ export const space = {
   "16": "1.6rem",
   "2": "0.2rem",
   "20": "2.0rem",
-  "208": "20.8rem",
+  "232": "23.2rem",
   "24": "2.4rem",
   "32": "3.2rem",
   "4": "0.4rem",
